+++ NEWS.md
@@ -1,6 +1,13 @@
 # MazamaTimeSeries 0.0.2
 
 * Added basic utility functions for `sts` and `mts` objects.
+* Added the following `sts` functions:
+  - `sts_filter()`
+  - `sts_filterDate()`
+  - `sts_filterDatetime()`
+  - `sts_join()`
+  - `sts_toTidyDF()`
+  - `sts_trimDate()`
 
 # MazamaTimeSeries 0.0.1